+++ changelog.md
@@ -7,6 +7,7 @@ you have backed up your save.**
 
 ## Changelog 0.10.7
 
+* Implemented part smelting
 * Added template and material resource overrides
 * Fixed crimson and warped wood not having attack damage
 * Fixed all basic recipes being changed to use the minecraft namespace for wood variants